django_celery_task/requirements.txt
celery
redis
django_celery_results